fix msvc warning 4244

+++ test/performance_trans100_fullspeed_test.cpp
@@ -67,7 +67,8 @@ class FullSpeed100ActorTransTest : public myframe::Actor {
         LOG(INFO) << "100 actor fullspeed trans msg avg(cnt/sec): " << avg;
         std::sort(msg_cnt_per_sec_list_.begin(), msg_cnt_per_sec_list_.end());
         LOG(INFO) << "100 actor fullspeed trans msg 99(cnt/sec): "
-          << msg_cnt_per_sec_list_[msg_cnt_per_sec_list_.size() * 0.99];
+          << msg_cnt_per_sec_list_[
+              static_cast<size_t>(msg_cnt_per_sec_list_.size() * 0.99)];
         GetApp()->Quit();
       }
     }
